Adjusted namespace for CombinedFamilyInfoFormatter
